agents/ingestion-agent/shared/requirements.txt
# Core shared dependencies
grpcio>=1.60.0
google-cloud-firestore>=2.13.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.14.0
google-genai>=0.3.0
google-cloud-pubsub>=2.18.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# Encryption & Security
cryptography>=41.0.0

# Data Structures
dataclasses-json>=0.6.0
pydantic>=2.9.0
typing-extensions>=4.12.0
